# apps/api/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lumina API",
    "description": "Lumina 营销智能助手统一 API。涵盖意图识别、Skill 调用、Marketing Orchestra 编排、Token 用量查询、Demo 工作台等能力层接口。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/usage/stats": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "统一查询用户 token 用量",
        "description": "统一查询用户 token 用量。\n\n- 仅传 user_id：返回累计摘要（summary）\n- 同时传 start_date + end_date：返回 summary + 按天明细（daily）",
        "operationId": "usage_stats_api_v1_usage_stats_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "用户唯一标识",
              "title": "User Id"
            },
            "description": "用户唯一标识"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "查询起始日期（YYYY-MM-DD），与 end_date 同时提供时返回按天明细",
              "title": "Start Date"
            },
            "description": "查询起始日期（YYYY-MM-DD），与 end_date 同时提供时返回按天明细"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "查询结束日期（YYYY-MM-DD），与 start_date 同时提供时返回按天明细",
              "title": "End Date"
            },
            "description": "查询结束日期（YYYY-MM-DD），与 start_date 同时提供时返回按天明细"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usage/summary": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "获取用户累计 token 用量摘要",
        "description": "获取用户累计 token 用量摘要。（兼容接口，推荐使用 /stats）",
        "operationId": "usage_summary_api_v1_usage_summary_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "用户唯一标识",
              "title": "User Id"
            },
            "description": "用户唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usage/daily": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "按天汇总用户的 token 用量",
        "description": "按天汇总用户的 token 用量。（兼容接口，推荐使用 /stats）",
        "operationId": "usage_daily_api_v1_usage_daily_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "用户唯一标识",
              "title": "User Id"
            },
            "description": "用户唯一标识"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "查询起始日期（YYYY-MM-DD）",
              "title": "Start Date"
            },
            "description": "查询起始日期（YYYY-MM-DD）"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "查询结束日期（YYYY-MM-DD）",
              "title": "End Date"
            },
            "description": "查询结束日期（YYYY-MM-DD）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{export_id}": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "获取 HTML 导出文件",
        "description": "通过 export_id 获取 HTML 导出文件。\n\nexport_id 格式支持两种：\n1. UUID 格式：从数据库查询文件路径\n2. 相对路径格式：直接拼接 static/content/ 路径",
        "operationId": "get_export_api_v1_exports__export_id__get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UsageResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码，0 表示成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "状态描述"
          },
          "data": {
            "title": "Data",
            "description": "返回数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "UsageResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}